feat: enhance gateway with observability, resilience, and API documentation
- Add Prometheus metrics collection and /metrics endpoint
- Implement circuit breaker and retry logic for Fluss connectivity
- Add deep health check with ?deep=true parameter
- Integrate OpenAPI/Swagger UI at /swagger-ui
- Add comprehensive API documentation with utoipa
- Enhance health endpoint to report Fluss connectivity status
- Add request duration and error rate metrics
- Improve test coverage with resilience and metrics unit tests

All 26 unit tests and 19 integration tests pass successfully.

+++ src/config.rs
@@ -71,7 +71,7 @@ impl Default for AuthConfig {
 
 #[derive(Clone, Debug)]
 pub struct PoolConfig {
-    pub max_connections: u64,
+    pub max_connections: u32,
     pub idle_timeout_secs: u64,
 }
 
@@ -129,7 +129,7 @@ struct FileAuthConfig {
 
 #[derive(Deserialize, Default)]
 struct FilePoolConfig {
-    max_connections: Option<u64>,
+    max_connections: Option<u32>,
     idle_timeout_secs: Option<u64>,
 }
